It's becoming clear that we don't understand b64.

+++ src/b64/b64.h
@@ -9,7 +9,7 @@
 
 /* #define DEBUG_B64 */
 
-void b64_encode( BYTE* indata, int32_t indata_len, bstring outstring, int linesz );
+void b64_encode( BYTE* indata, size_t indata_len, bstring outstring, size_t linesz );
 BYTE* b64_decode( size_t* outdata_len, bstring instring );
 
 #endif /* B64_H */
